redid adjustment for focuser slewrate after upstream merge

+++ src/telescope/focuser/local/Focuser.cpp
@@ -18,7 +18,7 @@ typedef struct FocuserConfiguration {
   bool present;
   int32_t  homeDefault;
   uint16_t slewRateDesired;
-  uint8_t  slewRateMinimum;
+  uint16_t  slewRateMinimum;
   float    accelerationTime;
   float    rapidStopTime;
   bool     powerDown;
